Refactored header(separated searchbar); next task is handle matchhistory call upon render on the summonerpage component
