+++ frontend/src/App.js
@@ -3,12 +3,8 @@ import './App.css';
 import { CssBaseline } from '@material-ui/core';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
 import { createMuiTheme, ThemeProvider , responsiveFontSizes } from '@material-ui/core/styles';
-
-// import logo from './logo.svg';
-
 /* Static Components */
 import Header from './components/Header'
-
 /* Dynamic Components */
 import HomePage from './templates/HomePage'
 import SummonerPage from './templates/SummonerPage'
