+++ frontend/src/templates/SummonerPage.js
@@ -4,12 +4,6 @@ import { Grid } from '@material-ui/core';
 import SummonerInfo from '../components/SummonerInfo';
 import MatchHistoryCard from '../components/MatchHistoryCard';
 import MatchHistoryTabSelector  from '../components/MatchHistoryTabSelector';
-import {
-  useParams
-} from "react-router-dom";
-
-import MatchSummaryCard from '../components/MatchSummaryCard';
-import { useState } from 'react';
 
 const useStyles =  makeStyles((theme) =>({
   root: {
@@ -17,37 +11,11 @@ const useStyles =  makeStyles((theme) =>({
   },
 }));
 
-function SummonerPage(props) {
+function SummonerPage() {
   const classes = useStyles();
-  let { summoner, server} = useParams();
-  console.log(summoner, server);
-  
-  const [name, setName] = useState(summoner);
-  const [profileIconId, setIcon] = useState();
-    
-  setTimeout(() => {
-    let url = new URL('http://localhost:8000/api/summonerinfo');
-    url.search = new URLSearchParams([['summoner', summoner], ['server', server]]).toString();
-    console.log(url);
-    
-    const f = fetch(url);
-
-    const p = f.then(r => r.json());
-
-    console.log(p);
-
-    p.then((data) => {
-      console.log(data);
-      setTimeout(() => setName(data.name), 0);
-      console.log(name);
-      setTimeout(() => setIcon(data.profileIconId), 0);
-      console.log(profileIconId);
-    })}, 0
-  );
-
   return (
     <div className={classes.root}>
-      <Grid container direction="column" spacing={3}>
+      <Grid container direction="column">
         <SummonerInfo /> 
         {/* TODO: Setup <Switch> Routing to render depending on "All"   "Ranked"   "Normal" */}
         <MatchHistoryTabSelector /> {/* TODO: Need to refactor */}
